Fix use of statvfs on 32-bit platforms
A long type (4 bytes) cannot contain the available number of inodes and
like on 32-bit architectures with large file support enabled.  A long
long type is needed (8 bytes).

see #296

+++ storage/timecaf/caf.h
@@ -140,7 +140,7 @@ extern int CAFStatArticle(char *path, ARTNUM art, struct stat *st);
 */
 extern int CAFOpenReadTOC(char *cfpath, CAFHEADER *ch, CAFTOCENT **tocpp);
 extern int CAFReadHeader(int fd, CAFHEADER *h);
-extern off_t CAFRoundOffsetUp(off_t offt, unsigned int bsize);
+extern off_t CAFRoundOffsetUp(off_t offt, unsigned long int bsize);
 extern CAFBITMAP *CAFReadFreeBM(int fd, CAFHEADER *h);
 extern void CAFDisposeBitmap(CAFBITMAP *cbm);
